Time: 2449 ms (5.03%) | Memory: 7.6 MB (99.44%) - LeetSync

+++ 367-valid-perfect-square/valid-perfect-square.cpp
@@ -1,7 +1,7 @@
 class Solution {
 public:
     bool isPerfectSquare(int num) {
-        for(long long i=1;i<=num; i++){
+        for(long i=1;i<=num; i++){
             if(i*i==num){
             return true;
         }
